Chapter two.
Created chapter two, the literature review chapter.
Added subsection Prisoner's Dilemma. A section that explains what
the PD and the Iterated PD is.
Second subsection is the Tournament. It's a small review on the
first tournments conducted by Axelrod.
Third one, Spatial Tournaments. A small introduction on spatial structure
tournaments and on the work done by some authors.
Finally, a small introduction on Axelrod library and how it can be used
for the goals of this dissertation.