Adressing VCS and TDD.
A paragraph on unit testing has been written, as well as on
version control system.

For version cotrol system I have added url with the library and
explained the process of submiting. How we use pull request and
contributors review them. An introducing sentence has been added for
the next section.

Lastly, I have gone back to the start and altered the introduction.